+++ noodles-bam/src/reader/records.rs
@@ -5,13 +5,13 @@ use crate::Record;
 
 /// An iterator over lazily-evalulated records of a BAM reader.
 ///
-/// This is created by calling [`Reader::lazy_records`].
-pub struct LazyRecords<'a, R> {
+/// This is created by calling [`Reader::records`].
+pub struct Records<'a, R> {
     reader: &'a mut Reader<R>,
     record: Record,
 }
 
-impl<'a, R> LazyRecords<'a, R>
+impl<'a, R> Records<'a, R>
 where
     R: Read,
 {
@@ -23,14 +23,14 @@ where
     }
 }
 
-impl<'a, R> Iterator for LazyRecords<'a, R>
+impl<'a, R> Iterator for Records<'a, R>
 where
     R: Read,
 {
     type Item = io::Result<Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match self.reader.read_lazy_record(&mut self.record) {
+        match self.reader.read_record(&mut self.record) {
             Ok(0) => None,
             Ok(_) => Some(Ok(self.record.clone())),
             Err(e) => Some(Err(e)),